sahayak-coordinator/requirements.txt
# Google Cloud Dependencies
google-cloud-firestore==2.13.1
google-generativeai==0.3.2
google-cloud-storage==2.10.0
google-auth==2.23.4

# MCP Protocol (placeholder - will use actual ADK MCP implementation)
# mcp-client==1.0.0
# mcp-server==1.0.0

# Web Framework for API endpoints
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Async support
aiohttp==3.9.1

# OCR and Image Processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0